refactor: CouponManagement 추가
