+++ src/refactoring/components/admin/CouponManagement/index.ts
@@ -1,0 +1,1 @@
+export { CouponManagement } from './CouponManagement';
